S9/deployment/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.5.1%2Bcpu-cp38-cp38-linux_x86_64.whl
torchtext==0.3.1
requests_toolbelt
